Move Messaging types to new messaging.ts file + some renames
